增加文件存在判断

+++ backend/deploy.sh
@@ -2,4 +2,8 @@ git reset --hard
 git pull
 rm -f Package.resolved
 swift build -c release
-cp -r .build/release/CanaryBackend release/canary
+if [ -f ".build/release/CanaryBackend" ];then
+    cp -r .build/release/CanaryBackend release/canary
+  else  
+    cp -r .build/debug/CanaryBackend release/canary
+fi
